requirements.txt
alembic==1.0.11.dev0
asn1crypto==0.24.0
attrs==18.2.0
autobahn==17.10.1
Automat==0.6.0
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.7
beautifulsoup4==4.8.0
blinker==1.4
bottle==0.12.15
cbor==1.0.0
certifi==2018.8.24
Chameleon==2.24
chardet==3.0.4
CherryTree==0.38.8
Click==7.0
colorama==0.3.7
configparser==3.5.0b2
constantly==15.1.0
cryptography==2.6.1
cssselect==1.1.0
Cython==0.29.14
dataclasses==0.6
dbus-python==1.2.12
decorator==4.3.0
deprecation==2.0.6
dnspython==1.16.0
EditorConfig==0.12.1
Elixir==0.7.1
enum34==1.1.6
feedparser==5.2.1
filedepot==0.5.2
filetype==1.0.7
Flask==1.1.1
Flask-Classful==0.14.1
Flask-KVSession==0.6.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Restless==0.17.0
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
FormEncode==1.3.0
future==0.16.0
futures==3.3.0
html2text==2019.8.11
html5lib==1.0.1
hupper==1.5
hyperlink==17.3.1
idna==2.6
impacket==0.9.20
incremental==16.10.1
ipaddress==1.0.17
IPy==0.83
ipython==5.8.0
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.10.1
jsbeautifier==1.6.4
lxml==4.4.1
lz4==2.1.10+dfsg
Mako==1.0.7
MarkupSafe==1.1.0
marshmallow==3.0.0b3
marshmallow-sqlalchemy==0.14.1
mimerender==0.6.0
netaddr==0.7.19
nplusone==1.0.0
numpy==1.16.5
olefile==0.46
packaging==19.1
paramiko==2.6.0
passlib==1.7.1
Paste==3.2.2
pathlib2==2.3.5
peepdf==0.4.2
pefile==2019.4.18
pexpect==4.6.0
pickleshare==0.7.5
Pillow==6.2.0
plaster==1.0
prettytable==0.7.2
prompt-toolkit==1.0.15
psycopg2==2.8.4
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycrypto==2.6.1
pycryptodomex==3.6.1
pydot==1.4.1
pyenchant==2.0.0
pygame==1.9.6
Pygments==2.3.1
PyGObject==3.36.0
pyinotify==0.9.6
pylibemu==0.3.3
pymssql==2.1.4
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.2
pypcapfile==0.12.1
pypng==0.0.20
PyQRCode==1.2.1
pyquery==1.2.9
PySocks==1.6.8
python-dateutil==2.7.3
python-editor==1.0.3
python-magic==0.4.16
python-mimeparse==1.6.0
python-openid==2.2.5
python-snappy==0.5.3
python3-nmap==1.4.8
pythonaes==1.0
PyTrie==0.2
pytube3==9.5.5
pytz==2019.3
PyV8==1.0.dev0
PyYAML==5.1.2
qrcode==6.1
repoze.lru==0.7
requests==2.21.0
scandir==1.10.0
scapy==2.4.3
scgi==1.13
selenium==3.14.1
service-identity==18.1.0
simplegeneric==0.8.1
simplejson==3.16.0
simplekv==0.13.0
singledispatch==3.4.0.3
sip==4.19.19
six==1.13.0
soupsieve==1.9.4
speaklater==1.3
sphinx-rtd-theme==0.5.0
SQLAlchemy==1.3.10
sqlalchemy-schemadisplay==1.3
sqlparse==0.2.4
sslstrip==0.9
syslog-rfc5424-formatter==1.2.2
Tempita==0.5.2
tinytag==1.4.0
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.3
translationstring==1.3
trollius==2.0.1
Twisted==18.9.0
txaio==2.10.0
typing==3.7.4.3
typing-extensions==3.7.4.3
u-msgpack-python==2.1
Unidecode==1.1.1
urllib3==1.24.1
uTidylib==0.5
venusian==1.2.0
waitress==1.2.0b2
wcwidth==0.1.7
webencodings==0.5.1
WebOb==1.8.5
websocket-client==0.53.0
Werkzeug==0.16.0
wifite==2.2.5
wsaccel==0.6.2
WTForms==2.2.1
zenmap==6.47
zope.component==4.3.0
zope.deprecation==4.4.0
zope.event==4.2.0
zope.hookable==4.0.4
zope.interface==4.6.0
